olaf_python/requirements.txt
cffi==1.17.1
cryptography==43.0.1
pycparser==2.22
rel==0.4.9.19
websocket-client==1.8.0
